Add files via upload
This inverted index omits "and","but", ...  and supports AND and OR queries. We only need to put our txt files (our books, docs,... ) in a folder named Document and put it in the same directory as the python code is  running, query.txt file should also be placed in the same folder.